Added static casts

+++ Source/AliveLib/Mudokon.cpp
@@ -1008,8 +1008,8 @@ void Mudokon::vUpdate_4757A0()
                 vGetBoundingRect_424FD0(&bRect, 1);
 
                 vOnCollisionWith_424EE0(
-                    { bRect.x, bRect.y + 5 },
-                    { bRect.w, bRect.h + 5 },
+                    { bRect.x, static_cast<short>(bRect.y + 5) },
+                    { bRect.w, static_cast<short>(bRect.h + 5) },
                     ObjList_5C1B78,
                     1,
                     (TCollisionCallBack)&BaseAliveGameObject::OnTrapDoorIntersection_408BA0);
